% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrange_by_genomic_position}
\alias{arrange_by_genomic_position}
\title{Arrange a data frame by genomic position based on a \code{locus} column}
\usage{
arrange_by_genomic_position(df)
}
\arguments{
\item{df}{A tibble or data.frame with a \code{locus} column in the format \code{"{chr}:{start}-{end}"}
where \code{{chr}} is 1-22, X, or Y and where \code{{start}} and \code{{end}} are genomic coordinates.}
}
\value{
Input data frame arranged by genomic position
}
\description{
Arrange a data frame by genomic position based on a \code{locus} column
}
\keyword{internal}
