% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal_noise.R
\name{calc.signal.noise}
\alias{calc.signal.noise}
\title{Calculate the signal and noise profile of a sample}
\usage{
calc.signal.noise(list, pif)
}
\arguments{
\item{list}{A numeric vector representing the signal values for a sample}

\item{pif}{A data frame containing the positional information for each locus, including chromosome}
}
\description{
This function calculates the signal and noise profile of a sample. The signal is defined
as the standard deviation of the median signal of each arm (grouped into autosomes, X, and Y chromosomes),
and the noise is defined as the median of the absolute values of the vector of
consecutive differences between each signal.
}
\keyword{internal}
