% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cbs.R
\name{convert_cbs}
\alias{convert_cbs}
\title{Converts CBS output to signal format}
\usage{
convert_cbs(cbs_out, sif_df, tsig_df)
}
\arguments{
\item{cbs_out}{The output from \code{\link{run_cbs}} (i.e. list of outputs from \code{\link[DNAcopy]{segment}}).}

\item{sif_df}{Tibble or filepath to a text file containing sample metadata}

\item{tsig_df}{Tibble or filepath to a text file containing the tumor signal matrix}
}
\value{
Tumor signal matrix in the same format as the input tumor signal matrix.
}
\description{
Converts output of \code{\link{run_cbs}} into the format of the
signal files (normal and tumor) used elsewhere in this package. Fills female Y chromosome values with NAs.
}
\keyword{internal}
