% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_nsig_df}
\alias{example_nsig_df}
\title{Example normal signal matrix file}
\format{
A tibble with 4,800 rows and 11 columns:
\describe{
\item{locus}{Locus in the format \code{"{chr}:{start}-{end}"} where \code{{chr}} is 1-22, X, or Y}
\item{sample.id}{One column for each normal sample containing log2(RCN) values}
}
}
\usage{
example_nsig_df
}
\description{
A tibble containing log2 relative copy number values for normal samples in a small test dataset
}
\keyword{datasets}
