% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pif}
\alias{example_pif}
\title{Example Probe Information File}
\format{
A tibble with 5 columns:
\describe{
\item{locus}{Locus in the format \code{"{chr}:{start}-{end}"} where \code{{chr}} is 1-22, X, or Y}
\item{chr}{Chromosome of the probe}
\item{start}{Start position of the probe}
\item{end}{End position of the probe}
\item{arm}{Arm of the chromosome (p or q)}
}
}
\usage{
example_pif
}
\description{
A tibble containing probe information for a small test dataset
}
\keyword{datasets}
