% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sif}
\alias{example_sif}
\title{Example Sample Information File}
\format{
A tibble with 4 columns:
\describe{
\item{sample.id}{Sample ID}
\item{patient.id}{Patient ID}
\item{sex}{One of "male" or "female"}
\item{type}{One of "normal" or "tumor"}
}
}
\usage{
example_sif
}
\description{
A tibble containing sample information for a small test dataset
}
\keyword{datasets}
