% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tsig_df}
\alias{example_tsig_df}
\title{Example tumor signal matrix file}
\format{
Analagous to \code{example_nsig_df}
}
\usage{
example_tsig_df
}
\description{
A tibble containing log2 relative copy number values for tumor samples in a small test dataset
}
\keyword{datasets}
