% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_blacklist.R
\name{filter_blacklist}
\alias{filter_blacklist}
\title{Filter out common germline copy-number variants}
\usage{
filter_blacklist(
  probe_info,
  blacklist = "hg38",
  start_col = "flanking.start",
  end_col = "flanking.end"
)
}
\arguments{
\item{probe_info}{Tibble or filepath to a text file containing probe information.
Should have a "locus" column of the following form: \code{"{chr}:{start}-{end}"} where \code{{chr}} is 1-22, X, or Y
and where \code{{start}} and \code{{end}} are genomic coordinates}

\item{blacklist}{By default \code{hg38}. If the data in \code{probe_info} is aligned to \code{hg19}, set this parameter to
\code{"hg19"}. A dataframe or filepath can also be supplied as a parameter for custom filtering.
Blacklist should have a \code{chr} column with chromosome names (1-22, X, or Y) and columns for start and end coordinates.}

\item{start_col}{The name of the column in the blacklist file to use as the start coordinate for filtering}

\item{end_col}{The name of the column in the blacklist file to use as the end coordinate for filtering}
}
\value{
Input \code{probe_info} dataframe with rows overlapping the blacklist removed
}
\description{
Filter out common germline copy-number variants
}
