% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_blacklist}
\alias{hg38_blacklist}
\title{Germline Copy Number Variants Blacklist Data Version hg38}
\format{
\subsection{\code{hg38_blacklist}}{

A data frame with 7,285 rows and 6 columns:
\describe{
\item{id}{Unique ID for CNV}
\item{chr}{Chromosome of CNV}
\item{start}{CNV start location}
\item{end}{CNV end location}
\item{flanking.start}{Start location of region flanking CNV}
\item{flanking.end}{End location of region flanking CNV}
}
}
}
\source{
\url{https://github.com/beroukhim-lab/gistic_reference/tree/main}
}
\usage{
hg38_blacklist
}
\description{
An updated CNV blacklist, used for GISTIC, derived from a SNP6 array analysis
with Cancer Gene Census loci removed
}
\keyword{datasets}
