% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{linear_transformation}
\alias{linear_transformation}
\title{Apply linear transformation to male chromosome X and Y signals}
\usage{
linear_transformation(sif, n.df, t.df)
}
\arguments{
\item{sif}{Tibble or data.frame containing sample metadata.}

\item{n.df}{Tibble or data.frame containing the normal signal matrix. Rownames
should be locus names in the format \code{"{chr}:{start}-{end}"} where \code{{chr}} is 1-22, X, or Y
and where \code{{start}} and \code{{end}} are genomic coordinates}

\item{t.df}{Tibble or data.frame containing the tumor signal matrix. Rownames
should match \code{n.df}.}
}
\value{
A list with two data frames, both having a \code{locus} column:
\describe{
\item{n.df: }{Normal signal matrix with male chromosome X and Y signals linearly transformed}
\item{t.df: }{Tumor signal matrix with male chromosome X signals linearly transformed}
}
}
\description{
Applies linear transformation to male normal sample chromosome X and Y signals.
Also applies linear transformation to male tumor sample chromosome X signals.
}
\keyword{internal}
