% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_factor_importance.R
\name{plot_latent_factor_importance}
\alias{plot_latent_factor_importance}
\title{Plot importance of latent factors}
\usage{
plot_latent_factor_importance(sif_df, nsig_df, tsig_df, output_dir = NULL)
}
\arguments{
\item{sif_df}{Tibble or filepath to a text file containing sample metadata}

\item{nsig_df}{Tibble or filepath to a text file containing the normal signal matrix}

\item{tsig_df}{Tibble or filepath to a text file containing the tumor signal matrix}

\item{output_dir}{Directory to save the plots. If \code{NULL}, the plots will be printed to the screen.}
}
\value{
(Invisibly) A list of ggplot objects. Component \code{all} contains the plot of all latent factors.
Component \code{top_100} contains the plot of the top 100 latent factors (if there are at least 100 latent factors).
}
\description{
Plot importance of latent factors
}
