% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNormalTransformation.R
\name{plot_normal_transformation}
\alias{plot_normal_transformation}
\title{Plots the signal distribution of Chromosome X after the linear transformation done by \code{transform_normals()}}
\usage{
plot_normal_transformation(sif_filepath, ndf_filepath, ndf_transformed)
}
\arguments{
\item{sif_filepath}{The filepath for the sample information file used to create \code{ndf_transformed}}

\item{ndf_filepath}{The filepath for the normal signal matrix file used to create \code{ndf_transformed}}

\item{ndf_transformed}{The returned transformed dataframe from \code{transform_normals()}}
}
\value{
A dataframe of the plotted data
}
\description{
Plots the signal distribution of Chromosome X after the linear transformation done by \code{transform_normals()}
}
