% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_x_and_y_signal_transformation.R
\name{plot_x_and_y_signal_transformation}
\alias{plot_x_and_y_signal_transformation}
\title{Plot Chromosome X and Y signals before and after linear transformation}
\usage{
plot_x_and_y_signal_transformation(
  sif,
  n.df,
  t.df,
  n.df.transformed,
  t.df.transformed,
  output_dir = NULL
)
}
\arguments{
\item{sif}{Tibble or data.frame containing sample metadata.}

\item{n.df}{Tibble or data.frame containing the normal signal matrix. Rownames
should be locus names in the format \code{"{chr}:{start}-{end}"} where \code{{chr}} is 1-22, X, or Y
and where \code{{start}} and \code{{end}} are genomic coordinates}

\item{t.df}{Tibble or data.frame containing the tumor signal matrix. Rownames
should match \code{n.df}.}

\item{n.df.transformed}{Tibble or data.frame containing the normal signal matrix after linear transformation.
Output of \link{linear_transformation}.}

\item{t.df.transformed}{Tibble or data.frame containing the tumor signal matrix after linear transformation.
Output of \link{linear_transformation}.}

\item{output_dir}{Directory to save the plots. If \code{NULL}, the plots will be printed to the screen.}
}
\value{
Nothing. Generates and saves or prints plots.
}
\description{
Plot Chromosome X and Y signals before and after linear transformation
}
\keyword{internal}
