% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_and_format_input}
\alias{read_and_format_input}
\title{Read in a data frame or filepath with a \code{locus} column and arrange it in
order of increasing genomic position (1-22, X, Y; within each chromosome, by start position).}
\usage{
read_and_format_input(input, locus_to_rownames = FALSE)
}
\arguments{
\item{input}{A tibble or a filepath to a text file containing the data. Must have a \code{locus} column
in the format \code{"{chr}:{start}-{end}"} where \code{{chr}} is 1-22, X, or Y and where \code{{start}} and \code{{end}} are genomic
coordinates.}

\item{locus_to_rownames}{If \code{TRUE}, the \code{locus} column will be converted to rownames.}
}
\value{
Input data frame arranged by genomic position
}
\description{
Read in a data frame or filepath with a \code{locus} column and arrange it in
order of increasing genomic position (1-22, X, Y; within each chromosome, by start position).
}
\keyword{internal}
