% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtangentXY-package.R
\docType{package}
\name{RtangentXY-package}
\alias{RtangentXY}
\alias{RtangentXY-package}
\title{RtangentXY: TangentXY SCNA Signal Normalization}
\description{
This is the package implementation of TangentXY, a sex-informed copy number inference algorithm which extends Tangent, as published in 2022 (PMID: 36040167).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beroukhim-lab.github.io/RtangentXY/}
}

}
\author{
\strong{Maintainer}: Greg Raskind \email{graskind@umich.edu} (\href{https://orcid.org/0000-0002-9940-1984}{ORCID})

Authors:
\itemize{
  \item Kei Enomoto \email{Kei_Enomoto@dfci.harvard.edu}
  \item Siyun Lee \email{siyun@broadinstitute.org} (\href{https://orcid.org/0009-0005-5130-5212}{ORCID})
}

Other contributors:
\itemize{
  \item Rameen Beroukhim \email{Rameen_Beroukhim@dfci.harvard.edu} [copyright holder, funder]
}

}
\keyword{internal}
