% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cbs.R
\name{run_cbs}
\alias{run_cbs}
\title{Run CBS segmentation on normalized tumor signals}
\usage{
run_cbs(tnorm_df, alpha = 0.005, min_width = 3, n_cores = 1)
}
\arguments{
\item{tnorm_df}{Normalized tumor signal from \code{\link{run_tangent}}.}

\item{alpha}{Passed to \code{\link[DNAcopy]{segment}}.}

\item{min_width}{Passed to \code{\link[DNAcopy]{segment}}.}

\item{n_cores}{Number of cores to use for parallel processing.}
}
\value{
A named list with the outputs of the CBS algorithm. Each element corresponds to a sample.
}
\description{
Run CBS segmentation on normalized tumor signals
}
