% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPseudoTangent.R
\name{run_pseudotangent}
\alias{run_pseudotangent}
\title{Runs the Pseudotangent pipeline}
\usage{
run_pseudotangent(
  sif_filepath,
  ndf_filepath,
  tdf_filepath,
  n_latent_init,
  n_latent_tsvd
)
}
\arguments{
\item{sif_filepath}{The filepath for the sample information file}

\item{ndf_filepath}{The filepath for the normal signal matrix file}

\item{tdf_filepath}{The filepath for the tumor signal matrix file}

\item{n_latent_init}{An integer representing the number of latent factors to reconstruct normal subspace}

\item{n_latent_tsvd}{An integer representing the number of latent factors for each of the runs}
}
\value{
A normalized tumor signal matrix
}
\description{
Pseudotangent also isolates the tumor signals using the Tangent algorithm,
but it should be used when the set of normals are particularly non-representative.
}
