% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tangent.R
\name{run_tangent}
\alias{run_tangent}
\title{Run the TangentXY algorithm}
\usage{
run_tangent(
  sif_df,
  nsig_df,
  tsig_df,
  n_latent,
  make_plots = TRUE,
  output_dir = NULL
)
}
\arguments{
\item{sif_df}{Tibble or filepath to a text file containing sample metadata}

\item{nsig_df}{Tibble or filepath to a text file containing the normal signal matrix}

\item{tsig_df}{Tibble or filepath to a text file containing the tumor signal matrix}

\item{n_latent}{Number of latent factors to reconstruct normal subspace}

\item{make_plots}{If \code{TRUE}, generate plots of latent factor importance and effects of linear transformation}

\item{output_dir}{Directory to save the plots. If \code{NULL}, the plots will be printed to the screen.}
}
\value{
A normalized tumor signal matrix
}
\description{
Isolate tumor signal by removing systematic noise using a set of normal samples.
}
\examples{
res <- run_tangent(example_sif, example_nsig_df, example_tsig_df, 5)
head(res)

}
