% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cbs.R
\name{sample.cbs}
\alias{sample.cbs}
\title{Run CBS segmentation on a single sample}
\usage{
sample.cbs(sample, alpha, min.width)
}
\description{
Uses the start position as the probe location.
}
\keyword{internal}
