% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_segmentation.R
\name{save_segmentation}
\alias{save_segmentation}
\title{Save segmentation file}
\usage{
save_segmentation(seg_list, output_dir, filenames = NULL)
}
\arguments{
\item{seg_list}{The output of \code{\link{run_cbs}}, i.e. a list of DNAcopy objects}

\item{output_dir}{Directory where the segmentation files should be saved}

\item{filenames}{A character vector of the names to save each sample by. This should be the same length and order as \code{seg_list}}
}
\value{
1 if the file was successfully saved
}
\description{
Save each sample's segmentation file as a tab-delimited text file.
}
