% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalNoise.R
\name{signal_noise}
\alias{signal_noise}
\title{Calculate the signal and noise profile of each sample}
\usage{
signal_noise(signal_df, pif_df)
}
\arguments{
\item{signal_df}{The tibble of or the filepath for the tumor signal file (pre- or post-normalization)}

\item{pif_df}{The tibble of or the filepath for the probe information file}
}
\value{
A dataframe of the signal and noise profiles for each sample
}
\description{
This function calculates the signal and noise profile of each sample. In this
case, the signal is defined as the standard deviation of the median signal of
each arm (grouped into autosomes, X, and Y chromosomes), and the noise is
defined as the median of the absolute values of the vector of consecutive
differences between each signal.
}
