% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMatrix.R
\name{to_matrix}
\alias{to_matrix}
\title{Convert Read Counts to Matrix}
\usage{
to_matrix(...)
}
\arguments{
\item{...}{The tibble of or the filepath for the read counts}
}
\value{
A pre-processed read counts matrix
}
\description{
This takes in the filepath or tibble of GATK DepthOfCoverage or CollectReadCounts
outputs, checks if they contain the same loci, and preprocesses them for the
Tangent pipeline. Input either all tumor signals or normal signals to be used
for the analysis.
}
