% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformNormals.R
\name{transform_normals}
\alias{transform_normals}
\title{Transform Chromosome X signals}
\usage{
transform_normals(sif_filepath, ndf_filepath)
}
\arguments{
\item{sif_filepath}{The filepath for the sample information file}

\item{ndf_filepath}{The filepath for the normal signal matrix file}
}
\value{
A matrix with the male X chromosomes linearly transformed
}
\description{
Conduct linear transformation on the CN signals of the male X chromosomes.
}
