% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCBS.R
\name{convert_cbs}
\alias{convert_cbs}
\title{Converts CBS output to signal format}
\usage{
convert_cbs(cbs_out, sif_df, tsig_df)
}
\arguments{
\item{cbs_out}{The output from \code{runCBS()} (i.e. list of outputs from \code{DNAcopy::segment()})}

\item{sif_df}{The tibble of or the filepath for the sample information file}

\item{tsig_df}{The tibble of or the filepath for the tumor signal matrix file}
}
\value{
A normalized tumor signal matrix
}
\description{
This takes the output from the \code{runCBS()} function in this package (ie. the
output in the DNAcopy pakage format) and converts it into the format of the
signal files (normal and tumor) used elsewhere in this package. This makes
computation between the CBS outputs and signal files simple.
}
