% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBlacklist.R
\name{filter_blacklist}
\alias{filter_blacklist}
\title{Filter common germline CN variants}
\usage{
filter_blacklist(df, version = "hg38", col_range = NULL, cores = 2)
}
\arguments{
\item{df}{The tibble of or the filepath for the file to filter. Must contain a \code{locus}
column of the following form: \code{"{chr}:{start}-{end}"} where \code{{chr}} is 1-22 or X,
and where \code{{start}} and \code{{end}} are all integers}

\item{version}{By default \code{hg38}. If the data in \code{df} is aligned to \code{hg19}, set this parameter to
\code{"hg19"}. A dataframe or filepath can also be supplied as a parameter for custom filtering. Must contain a \code{chr}
column and columns with start and end CNV loci}

\item{col_range}{By default \code{NULL}, as the flanking regions of the blacklist files are selected.
If a custom version dataframe is supplied (or if the flanking regions are not desired), a numeric
vector can be supplied in this field of the form \code{c({start}, {end})}, where \code{{start}} and \code{{end}}
are integers corresponding to the desired start and end columns in \code{version}}

\item{cores}{The number of cores to parallelize with}
}
\value{
A normalized tumor signal matrix
}
\description{
Pseudotangent also isolates the tumor signals using the Tangent algorithm,
but it should be used when the set of normals are particularly non-representative.
NOTE: all signal file rows should be in order (ie. from 1-22, X, Y).
}
