% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19_blacklist}
\alias{hg19_blacklist}
\title{Germline Copy Number Variants Blacklist Data Version hg19}
\format{
A data frame with 7,454 rows and 6 columns:
\describe{
\item{id}{Unique ID for CNV}
\item{chr}{Chromosome of CNV}
\item{start}{CNV start location}
\item{end}{CNV end location}
\item{flanking.start}{Start location of region flanking CNV}
\item{flanking.end}{End location of region flanking CNV}
}
}
\source{
\url{ftp://ftp.broadinstitute.org/pub/GISTIC2.0/hg19_support/CNV.hg19.bypos.111213.txt}
}
\usage{
hg19_blacklist
}
\description{
A CNV blacklist, used for GISTIC, derived from a SNP6 array analysis
}
\keyword{datasets}
