% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformNormals.R
\name{linear_transformation}
\alias{linear_transformation}
\title{Transform Chromosome X signals}
\usage{
linear_transformation(sif_df, nsig_df, tsig_df)
}
\arguments{
\item{sif_df}{The tibble of or the filepath for the sample information file}

\item{nsig_df}{The tibble of or the filepath for the normal signal matrix file}
}
\value{
A matrix with the male X chromosomes linearly transformed
}
\description{
Conduct linear transformation on the CN signals of the male X chromosomes.
}
