% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLatentFactors.R
\name{plot_latent_factors}
\alias{plot_latent_factors}
\title{Plots the importance of latent factors  \code{run_svd()}}
\usage{
plot_latent_factors(n_svd, top_100 = FALSE)
}
\arguments{
\item{n_svd}{The returned matrix decomposition of \code{run_svd()}}

\item{top_100}{A boolean. Set to TRUE if interested in the top 100 latent factors. If set to TRUE and if there are fewer than 100 latent factors, an error will be thrown}
}
\value{
A boolean of whether the plot was successfully generated.
}
\description{
Plots the importance of latent factors  \code{run_svd()}
}
