% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_factor_importance.R
\name{plot_latent_factors}
\alias{plot_latent_factors}
\title{Plot importance of latent factors}
\usage{
plot_latent_factors(n.autox.svd, output_dir = NULL)
}
\arguments{
\item{n.autox.svd}{A list containing the SVD components (output of \link{run_svd})}
}
\value{
A list containing ggplot objects. If \code{output_dir} is specified, the plots will also be saved as .jpg files in the specified directory.
}
\description{
Plot singular values to visualize the importance of latent factors
}
\keyword{internal}
