% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNormalTransformation.R
\name{plot_normal_transformation}
\alias{plot_normal_transformation}
\title{Plot Chromosome X transformation}
\usage{
plot_normal_transformation(sif_df, nsig_df, ndf_transformed)
}
\arguments{
\item{sif_df}{The dataframe of or the filepath for the sample information file used to create \code{ndf_transformed}}

\item{nsig_df}{The dataframe of or the filepath for the normal signal matrix file used to create \code{ndf_transformed}}

\item{ndf_transformed}{The returned transformed dataframe from \code{transform_normals()}}
}
\value{
A dataframe of the plotted data
}
\description{
Plots the signal distribution of Chromosome X after the linear transformation
done by \code{transform_normals()}
}
