% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignalNoise.R
\name{plot_signal_noise}
\alias{plot_signal_noise}
\title{Plot signal and noise data}
\usage{
plot_signal_noise(sif_df, pif_df, nsig_df, tsig_df, n_latents, cores = 2)
}
\arguments{
\item{sif_df}{The dataframe of or the filepath for the sample information file}

\item{pif_df}{The dataframe of or the filepath for the probe information file}

\item{nsig_df}{The dataframe of or the filepath for the normal signal matrix file}

\item{tsig_df}{The dataframe of or the filepath for the tumor signal matrix file}

\item{n_latents}{A vector of latent factors to reconstruct the normal subspace}

\item{cores}{The number of parallel cores to use when calculating signal to noise. Default is set to 2 cores}
}
\value{
A matrix of the signal and noise data that is plotted with this function as well as the \code{run_tangent()} outputs. These are stored in a list under \code{signal_noise_df} and \code{tangent_out_df} respectively.
}
\description{
For each of the given number of latent factors, this function runs \code{run_tangent()} and plots the signal and noise data for each latent factor.
}
