% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignalNoise.R
\name{plot_signal_noise}
\alias{plot_signal_noise}
\title{Plot signal and noise data}
\usage{
plot_signal_noise(
  sif_filepath,
  pif_filepath,
  ndf_filepath,
  tdf_filepath,
  n_latents,
  cores = 2
)
}
\arguments{
\item{sif_filepath}{The filepath for the sample information file}

\item{pif_filepath}{The filepath for the probe information file}

\item{ndf_filepath}{The filepath for the normal signal matrix file}

\item{tdf_filepath}{The filepath for the tumor signal matrix file}

\item{n_latents}{A vector of latent factors to reconstruct the normal subspace}

\item{cores}{The number of parallel cores to use when calculating signal to noise. Default is set to 2 cores}
}
\value{
A matrix of the signal and noise data that is plotted with this function as well as the \code{run_tangent()} outputs. These are stored in a list under \code{signal_noise_df} and \code{tangent_out_df} respectively.
}
\description{
For each of the given number of latent factors, this function runs \code{run_tangent()} and plots the signal and noise data for each latent factor.
}
