% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCBS.R
\name{run_cbs}
\alias{run_cbs}
\title{Run CBS segmentation on normalized tumor signals}
\usage{
run_cbs(tnorm_df, data_type = "logratio", alpha = 0.005, min_width = 3)
}
\arguments{
\item{tnorm_df}{The \code{runTangent()} output dataframe. It should be in the same format as the tumor signal matrix file}

\item{data_type}{This is the \code{data.type} argument for \code{DNAcopy::CNA()}. Default is logratio, but can be \code{'binary'} if LOH data}

\item{alpha}{This is the \code{alpha} argument for \code{DNAcopy::segment()}. Default is 0.005}

\item{min_width}{This is the \code{min.width} argument for \code{DNAcopy::segment()}. Default is 3}
}
\value{
A named list with the outputs of the CBS algorithm. Each element is another list, returned by \code{DNAcopy::segment()}
}
\description{
The function \code{run_cbs} expects the output of \code{runTangent()} or an input of
the same format. This function expects the locus column to be in the format
chr:start-end (e.g. 1:500-9501). There are more arguments in \code{DNAcopy::segment}
that aren't modified in this function and package. In this case, the \code{R/runCBS.R}
file can be modified to include these arguments.
}
