% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pseudo_tangent.R
\name{run_pseudotangent}
\alias{run_pseudotangent}
\title{Run the PseudoTangentXY algorithm}
\usage{
run_pseudotangent(
  sif_df,
  nsig_df,
  tsig_df,
  n_latent_init,
  num_partitions,
  n_latent_part,
  cbs_a = 0.005,
  cbs_mw = 3,
  cbs_ncores = 1,
  partition_seed = 37,
  make_plots = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{sif_df}{Tibble or filepath to a text file containing sample metadata}

\item{nsig_df}{Tibble or filepath to a text file containing the normal signal matrix}

\item{tsig_df}{Tibble or filepath to a text file containing the tumor signal matrix}

\item{n_latent_init}{The number of latent factors to reconstruct the initial normal subspace}

\item{num_partitions}{The number of partitions to create in the pseudotangent pipeline. This
must be less than the number of tumors}

\item{n_latent_part}{The number of latent factors for each of the partition runs. This should not exceed the minimum number of normal/pseudonormal samples across all partitions.}

\item{cbs_a}{The alpha parameter for the function \code{\link{run_cbs}}}

\item{cbs_mw}{The minimum width parameter for the function \code{\link{run_cbs}}}

\item{cbs_ncores}{The number of cores to use for the function \code{\link{run_cbs}}}

\item{partition_seed}{The seed to set for reproducibility for the random partitioning}

\item{make_plots}{If \code{TRUE}, generate plots of latent factor importance and effects of linear transformation}

\item{output_dir}{Directory to save the plots. If \code{NULL}, the plots will be printed to the screen.}
}
\value{
A normalized tumor signal matrix
}
\description{
Pseudotangent also isolates the tumor signals using the Tangent algorithm,
but it should be used when the set of normals are particularly non-representative.
}
