% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPseudoTangent.R
\name{run_pseudotangent}
\alias{run_pseudotangent}
\title{Runs the Pseudotangent pipeline}
\usage{
run_pseudotangent(
  sif_df,
  nsig_df,
  tsig_df,
  n_latent_init,
  cbs_dt = "logratio",
  cbs_a = 0.005,
  cbs_mw = 3,
  num_partitions,
  partition_seed = 37,
  n_latent_part
)
}
\arguments{
\item{sif_df}{The tibble of or the filepath for the sample information file}

\item{nsig_df}{The tibble of or the filepath for the normal signal matrix file}

\item{tsig_df}{The tibble of or the filepath for the tumor signal matrix file}

\item{n_latent_init}{The number of latent factors to reconstruct the initial normal subspace}

\item{cbs_dt}{The data type parameter for the function \code{run_cbs()}}

\item{cbs_a}{The alpha parameter for the function \code{run_cbs()}}

\item{cbs_mw}{The minimum width parameter for the function \code{run_cbs()}}

\item{num_partitions}{The number of partitions to create in the pseudotangent pipeline. This
must be less than the number of tumors}

\item{partition_seed}{The seed to set for reproducibility for the random partitioning}

\item{n_latent_part}{The number of latent factors for each of the partition runs. This should not exceed the minimum number of normal/pseudonormal samples across all partitions.}
}
\value{
A normalized tumor signal matrix
}
\description{
Pseudotangent also isolates the tumor signals using the Tangent algorithm,
but it should be used when the set of normals are particularly non-representative.
NOTE: all signal file rows should be in order (ie. from 1-22, X, Y).
}
