% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_svd.R
\name{run_svd}
\alias{run_svd}
\title{Calculate SVD of linearly transformed normal signal matrix}
\usage{
run_svd(nt.df)
}
\arguments{
\item{nt.df}{A data frame containing the linearly transformed normal signal matrix (output of \code{\link{linear_transformation}})}
}
\value{
A list containing the SVD components
}
\description{
Calculate singular value decomposition (SVD) of the linearly transformed normal signal matrix
in preparation for input into the Tangent algorithm.
}
\keyword{internal}
