% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSVD.R
\name{run_svd}
\alias{run_svd}
\title{Conduct SVD}
\usage{
run_svd(sif_filepath, ndf_filepath)
}
\arguments{
\item{sif_filepath}{The filepath for the sample information file}

\item{ndf_filepath}{The filepath for the normal signal matrix file}
}
\value{
A matrix with the male X chromosomes linearly transformed
}
\description{
Conduct singular value decomposition (SVD) on the normal signals in
preparation for input into the Tangent algorithm
}
