% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTangent.R
\name{run_tangent}
\alias{run_tangent}
\title{Run the TangentXY algorithm}
\usage{
run_tangent(sif_filepath, ndf_filepath, tdf_filepath, n_latent)
}
\arguments{
\item{sif_filepath}{The filepath for the sample information file}

\item{ndf_filepath}{The filepath for the normal signal matrix file}

\item{tdf_filepath}{The filepath for the tumor signal matrix file}

\item{n_latent}{An integer representing the number of latent factors to reconstruct normal subspace}
}
\value{
A normalized tumor signal matrix
}
\description{
This function isolates isolates the tumor signal usng a set of normal signals by running the Tangent algorithm.
}
