% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSegmentation.R
\name{save_segmentation}
\alias{save_segmentation}
\title{Save segmentation file}
\usage{
save_segmentation(seg_list, dirpath, filenames = NULL)
}
\arguments{
\item{seg_list}{The output of \code{run_cbs()}, i.e. a list of DNAcopy objects}

\item{dirpath}{The path of the directory where the segmentation files should be saved}

\item{filenames}{A character vector of the names to save each sample by. This should be the same length and order as \code{seg_list}}
}
\value{
1 if the file was successfully saved
}
\description{
Save the output of CBS to a given filepath as a segmentation file.
}
