% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformNormals.R
\name{transform_normals}
\alias{transform_normals}
\title{Transform Chromosome X signals}
\usage{
transform_normals(sif_df, nsig_df)
}
\arguments{
\item{sif_df}{The dataframe of or the filepath for the sample information file}

\item{nsig_df}{The dataframe of or the filepath for the normal signal matrix file}
}
\value{
A matrix with the male X chromosomes linearly transformed
}
\description{
Conduct linear transformation on the CN signals of the male X chromosomes.
}
